requirements.txt
blinker==1.9.0
chardet==7.4.3
click==8.3.3
colorama==0.4.6
Flask==3.1.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.0.0
pillow==12.2.0
python-dotenv==1.0.1
reportlab==4.2.2
Werkzeug==3.1.8
